perf(InViewport): use intersection observer to calculate inviewport state
Replace custom logic to calculate `InViewport` status for native `IntersectionObserver` API to improve
performance and clean up code

BREAKING CHANGE: Removed `forRoot` method in module which is no longer required for `AppBrowserModule`.
Replaced with `forServer` method for `AppServerModule`. Removed debounce feature and rxjs dependancy to
leave implementation up to the consumer of the library. This reduces bundle size if debounce feature is not being used.
Updated inviewport classes to `sn-inviewport--in` and `sn-inviewport--out` to match SOON styleguide